requirements.txt
bs4
requests
requests_toolbelt
pillow
chardet
jieba
aiocqhttp
python-opencv
annoy